[WIP] Rewrite vm sandbox tests #248
