+++ vm-api/src/main/scala/pravda/vm/RuntimeException.scala
@@ -23,12 +23,3 @@ final case class RuntimeException(error: Error,
                                   finalState: FinalState,
                                   callStack: Seq[(Option[Address], Seq[Int])],
                                   lastPosition: Int)
-
-//object RuntimeException {
-//
-//  def constructStackTraceLine(metas: List[Meta], pos: Int): String = {
-//    val translatorMessage = metas.collectFirst { case TranslatorMark(mark) => mark }
-//    val sources = metas.collectFirst { case s: SourceMark                  => s.markString }
-//    s"${translatorMessage.getOrElse(s"program:$pos")}${sources.map(s => s" ($s)").getOrElse("")}"
-//  }
-//}
