+++ vm-api/src/main/scala/pravda/vm/tethys.scala
@@ -1,0 +1,3 @@
+package pravda.vm
+
+object tethys extends TethysInstances
